+++ setup/DB/create-database.sql
@@ -1,2 +1,1 @@
 CREATE DATABASE ska_dlm_meta WITH TEMPLATE = template0 ENCODING = 'UTF8' LOCALE = 'en_US.utf8';
-ALTER DATABASE ska_dlm_meta OWNER TO ska_dlm_admin;