+++ setup/DB/create-tables.sql
@@ -1,6 +1,18 @@
 --
--- ska_dlm_adminQL DDL for SKA Data Lifecycle Management DB setup
+-- SQL DDL for SKA Data Lifecycle Management DB setup
 --
+
+--
+-- lookup tables
+--
+
+CREATE TABLE IF NOT EXISTS location_facility (
+    id TEXT PRIMARY KEY
+);
+
+INSERT INTO location_facility (id)
+SELECT unnest(ARRAY['SRC', 'STFC', 'AWS', 'Google', 'Pawsey Centre', 'external', 'local'])
+ON CONFLICT DO NOTHING;
 
 --
 -- Table location
@@ -9,15 +21,14 @@
 CREATE TABLE IF NOT EXISTS location (
     location_id uuid DEFAULT gen_random_uuid() PRIMARY KEY,
     location_name varchar NOT NULL UNIQUE,
-    location_type varchar NOT NULL,
-    location_country varchar DEFAULT NULL,
+    location_type location_type NOT NULL,
+    location_country location_country DEFAULT NULL,
     location_city varchar DEFAULT NULL,
-    location_facility varchar DEFAULT NULL,
+    location_facility TEXT DEFAULT NULL REFERENCES location_facility(id),
     location_check_url varchar DEFAULT NULL,
     location_last_check TIMESTAMP without time zone DEFAULT NULL,
     location_date timestamp without time zone DEFAULT now()
 );
-ALTER TABLE location OWNER TO ska_dlm_admin;
 
 --
 -- Table storage
@@ -28,9 +39,9 @@ CREATE TABLE IF NOT EXISTS storage (
     location_id uuid NOT NULL,
     storage_name varchar NOT NULL UNIQUE,
     root_directory varchar DEFAULT NULL,
-    storage_type varchar NOT NULL,
-    storage_interface varchar NOT NULL,
-    storage_phase_level varchar DEFAULT 'GAS',
+    storage_type storage_type NOT NULL,
+    storage_interface storage_interface NOT NULL,
+    storage_phase phase_type DEFAULT 'GAS',
     storage_capacity BIGINT DEFAULT -1,
     storage_use_pct NUMERIC(3,1) DEFAULT 0.0,
     storage_permissions varchar DEFAULT 'RW',
@@ -47,7 +58,6 @@ CREATE TABLE IF NOT EXISTS storage (
       REFERENCES location(location_id)
       ON DELETE SET NULL
 );
-ALTER TABLE storage OWNER TO ska_dlm_admin;
 
 --
 -- Table storage_config holds a JSON version of the configuration
@@ -59,7 +69,7 @@ ALTER TABLE storage OWNER TO ska_dlm_admin;
 CREATE TABLE IF NOT EXISTS storage_config (
     config_id uuid DEFAULT gen_random_uuid() PRIMARY KEY,
     storage_id uuid NOT NULL,
-    config_type varchar DEFAULT 'rclone',
+    config_type config_type DEFAULT 'rclone',
     config json NOT NULL,
     config_date timestamp without time zone DEFAULT now(),
     CONSTRAINT fk_cfg_storage_id
@@ -67,7 +77,6 @@ CREATE TABLE IF NOT EXISTS storage_config (
       REFERENCES storage(storage_id)
       ON DELETE SET NULL
 );
-ALTER TABLE storage_config OWNER TO ska_dlm_admin;
 
 
 --
@@ -86,10 +95,11 @@ CREATE TABLE IF NOT EXISTS data_item (
     item_type varchar DEFAULT 'file',
     item_format varchar DEFAULT 'unknown',
     item_encoding varchar DEFAULT 'unknown',
-    item_mime_type varchar DEFAULT 'application/octet-stream',
+    item_mime_type mime_type DEFAULT 'application/octet-stream',
     item_level smallint DEFAULT -1,
-    item_phase varchar DEFAULT 'GAS',
-    item_state varchar DEFAULT 'INITIALIZED',
+    uid_phase phase_type DEFAULT 'GAS',
+    oid_phase phase_type DEFAULT 'GAS',
+    item_state item_state DEFAULT 'INITIALISED',
     UID_creation timestamp without time zone DEFAULT now(),
     OID_creation timestamp without time zone DEFAULT NULL,
     UID_expiration timestamp without time zone DEFAULT now() + time '24:00',
@@ -100,7 +110,7 @@ CREATE TABLE IF NOT EXISTS data_item (
     deleted boolean DEFAULT false,
     last_access timestamp without time zone,
     item_checksum varchar,
-    checksum_method varchar DEFAULT 'none',
+    checksum_method checksum_method DEFAULT 'none',
     last_check timestamp without time zone,
     item_owner varchar DEFAULT 'SKA',
     item_group varchar DEFAULT 'SKA',
@@ -111,13 +121,13 @@ CREATE TABLE IF NOT EXISTS data_item (
     compression_method varchar DEFAULT NULL,
     parents uuid DEFAULT NULL,
     children uuid DEFAULT NULL,
-    metadata json DEFAULT NULL,
+    metadata jsonb DEFAULT NULL,
     CONSTRAINT fk_storage
       FOREIGN KEY(storage_id)
       REFERENCES storage(storage_id)
       ON DELETE SET NULL
 );
-ALTER TABLE data_item OWNER TO ska_dlm_admin;
+
 CREATE INDEX IF NOT EXISTS idx_fk_storage_id ON data_item USING btree (storage_id);
 
 CREATE UNIQUE INDEX IF NOT EXISTS idx_unq_OID_UID_item_version ON data_item USING btree (OID, UID, item_version);
@@ -154,10 +164,9 @@ FOR EACH ROW EXECUTE PROCEDURE
 CREATE TABLE IF NOT EXISTS phase_change (
     phase_change_ID bigint GENERATED always as IDENTITY PRIMARY KEY,
     OID uuid NOT NULL,
-    requested_phase varchar DEFAULT 'GAS',
+    requested_phase phase_type DEFAULT 'GAS',
     request_creation timestamp without time zone DEFAULT now()
 );
-ALTER TABLE phase_change OWNER TO ska_dlm_admin;
 
 
 --
@@ -187,4 +196,4 @@ CREATE TABLE IF NOT EXISTS migration (
       REFERENCES storage(storage_id)
       ON DELETE SET NULL
 );
-ALTER TABLE migration OWNER TO ska_dlm_admin;
+