+++ src/ska_dlm/dlm_storage/dlm_storage_requests.py
@@ -98,18 +98,18 @@ def init_storage(
     location_id: str | None = None,
     location_name: str | None = None,
     storage_capacity: int = -1,
-    storage_phase_level: str = "GAS",
+    storage_phase: str = "GAS",
     rclone_config: JsonObjectOption = None,
 ) -> str:
     """
-    Initialize a new storage.
+    Initialise a new storage.
 
     Parameters
     ----------
     storage_name
         An organisation or owner name for the storage.
     storage_type
-        high level type of the storage, e.g. "disk", "s3"
+        high level type of the storage: 'filesystem', 'objectstore' or 'tape'
     storage_interface
         storage interface for rclone access, e.g. "posix", "s3"
     root_directory
@@ -120,7 +120,7 @@ def init_storage(
         a dlm registered location name
     storage_capacity
         reserved storage capacity in bytes
-    storage_phase_level
+    storage_phase
         one of "GAS", "LIQUID", "SOLID"
     rclone_config
         extra rclone values such as secrets required for connection
@@ -137,7 +137,7 @@ def init_storage(
         "storage_interface",
         "location_id",
         "storage_capacity",
-        "storage_phase_level",
+        "storage_phase",
         "root_directory",
     ]
     # TODO remove keys none values
@@ -181,7 +181,7 @@ def create_storage_config(
     storage_name
         the name of the storage for which the config is provided.
     config_type
-        default is rclone, but could be something else in the future.
+        default is 'rclone'. Alternative enums from ConfigType.
 
     Returns
     -------
@@ -208,10 +208,10 @@ def create_storage_config(
 
 
 @cli.command()
-@rest.get("/storage/get_storage_config", response_model=list[str])
+@rest.get("/storage/get_storage_config", response_model=list[dict])
 def get_storage_config(
     storage_id: str = "", storage_name: str = "", config_type: str = "rclone"
-) -> list[str]:
+) -> list[dict]:
     """Get the storage configuration entry for a particular storage backend.
 
     Parameters
@@ -225,7 +225,7 @@ def get_storage_config(
 
     Returns
     -------
-    list[str]
+    list[dict]
         list of configs as json
 
     Raises
@@ -395,16 +395,16 @@ def init_location(
     location_city: str = "",
     location_facility: str = "",
 ) -> str:
-    """Initialize a new storage location.
+    """Initialise a new storage location.
 
     Parameters
     ----------
     location_name
         the orgization or owner's name managing the storage location.
     location_type
-        the location type, e.g. "server"
+        the location type, e.g. "low-operational"
     location_country
-        the location country name
+        the location country name (AU, ZA or UK)
     location_city
         the location city name
     location_facility