+++ tests/integration/client/dlm_migration_client.py
@@ -22,7 +22,7 @@ def copy_data_item(
     Steps
     (1) get the current storage_id(s) of the item
     (2) convert one (first) storage_id to a configured rclone backend
-    (3) initialize the new item with the same OID on the new storage
+    (3) initialise the new item with the same OID on the new storage
     (4) use the rclone copy command to copy it to the new location
     (5) set the access path to the payload
     (6) set state to READY