+++ CHANGELOG.md
@@ -2,10 +2,22 @@
 
 ## Development
 
+=======
+### Changed
+
+* Changed data_item.metadata column type from json to jsonb.
+* Remove ALTER TABLE ownership statements.
+* Added a patch SQL script for v1.1.2
+
+### Added
+
+* List of location facilities, as a lookup table
+* ENUM definitions for location_type, location_country, config_type, storage_type, storage_interface, phase_type, item_state, checksum_method and mime_type.
+* Set up the Helm chart infrastructure to support database patch releases via versioned SQL files.
+
 ## 1.1.3
 
 * Add optional `root_path` to rclone config and rc commands
-
 
 ## 1.1.2